requirements.txt
streamlit
pandas
plotly
openpyxl
plotly-express
datetime
typing
json
sqlite3
st-supabase-connection